infernal_ext/requirements.txt
pyautogui
pygetwindow
pillow
python-telegram-bot